let providers manage their own configs
