+++ src/providers/mod.rs
@@ -1,10 +1,12 @@
 mod cloudflare;
 mod duckdns;
 mod namecheap;
+mod porkbun;
 use anyhow::Result;
-pub use cloudflare::CloudflareManager;
-pub use duckdns::DuckdnsManager;
-pub use namecheap::NamecheapManager;
+pub use cloudflare::{CloudflareConfig, CloudflareManager};
+pub use duckdns::{DuckDNSConfig, DuckdnsManager};
+pub use namecheap::{NamecheapConfig, NamecheapManager};
+pub use porkbun::{PorkbunConfig, PorkbunManager};
 
 pub trait DnsProvider {
     fn update(&self, ip: &str) -> Result<String>;
